fix and test metadata setting
